spark-worker/requirements.txt
pyspark==3.5.0
redis==5.0.1
pytesseract==0.3.10
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
